Rename non-gradle code-gen tasks to steps
