+++ tools/mc-js/src/test/java/io/spine/tools/mc/js/code/task/ResolveImportsTest.java
@@ -32,6 +32,7 @@ import io.spine.tools.fs.ExternalModule;
 import io.spine.tools.fs.ExternalModules;
 import io.spine.tools.js.fs.Directory;
 import io.spine.tools.mc.js.code.given.GivenProject;
+import io.spine.tools.mc.js.code.step.ResolveImports;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
